extract filter tree into own cfile
